requirements.txt
opencv-python==4.2.0.34
numpy>=1.16.0
pillow==6.2.0
scipy==1.1.0
wget==3.2
seaborn==0.9.0
easydict==1.9
grpcio>=1.24.3
tensorflow==2.0.0
